refactor create-component.js
